+++ rebuild.sh
@@ -41,6 +41,8 @@ fi
 # CMake
 #
 [ -d build ] || mkdir build
+#cmake -S . -B build/Debug -DCMAKE_BUILD_TYPE=Debug -DCMAKE_C_COMPILER=/usr/bin/clang -DCMAKE_CXX_COMPILER=/usr/bin/clang++
+#cmake -S . -B build/Release -DCMAKE_BUILD_TYPE=Release -DCMAKE_C_COMPILER=/usr/bin/clang -DCMAKE_CXX_COMPILER=/usr/bin/clang++
 cmake -S . -B build/Debug -DCMAKE_BUILD_TYPE=Debug
 cmake -S . -B build/Release -DCMAKE_BUILD_TYPE=Release
 touch build/Release/compile_commands.json